% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBclust-package.R
\docType{package}
\name{IBclust-package}
\alias{IBclust}
\alias{IBclust-package}
\title{IBclust: Information-Based Clustering}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements multiple variants of the Information Bottleneck ('IB') method for clustering datasets containing continuous, categorical (nominal/ordinal) and mixed-type variables. The package provides deterministic, agglomerative, generalized, and standard 'IB' clustering algorithms that preserve relevant information while forming interpretable clusters. The Deterministic Information Bottleneck is described in Costa et al. (2026) \doi{10.1016/j.patcog.2026.113580}. The standard 'IB' method originates from Tishby et al. (2000) \doi{10.48550/arXiv.physics/0004057}, the agglomerative variant from Slonim and Tishby (1999) \url{https://papers.nips.cc/paper/1651-agglomerative-information-bottleneck}, and the generalized 'IB' from Strouse and Schwab (2017) \doi{10.1162/NECO_a_00961}.
}
\author{
\strong{Maintainer}: Angelos Markos \email{amarkos@gmail.com}

Authors:
\itemize{
  \item Efthymios Costa
  \item Ioanna Papatsouma
}

}
