% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_hclust.R
\name{as.hclust.aibclust}
\alias{as.hclust.aibclust}
\title{Convert an aibclust object to hclust}
\usage{
\method{as.hclust}{aibclust}(x, ...)
}
\arguments{
\item{x}{An \code{aibclust} object.}

\item{...}{Ignored.}
}
\value{
An object of class \code{"hclust"}.
}
\description{
Enables use of standard hierarchical-clustering methods
(e.g., \code{\link[stats]{cutree}}, \code{\link[stats]{as.dendrogram}})
on AIBmix output.
}
